Fix #888: wrong volume free/total size on macOS

+++ src/nnn.c
@@ -4181,9 +4181,9 @@ static size_t get_fs_info(const char *path, bool type)
 		return 0;
 
 	if (type == CAPACITY)
-		return svb.f_blocks << ffs((int)(svb.f_frsize >> 1));
+		return (size_t)svb.f_blocks << ffs((int)(svb.f_frsize >> 1));
 
-	return svb.f_bavail << ffs((int)(svb.f_frsize >> 1));
+	return (size_t)svb.f_bavail << ffs((int)(svb.f_frsize >> 1));
 }
 
 /* List or extract archive */
